+++ src/components/EmptyBox/EmptyBox.tsx
@@ -1,20 +1,19 @@
-    import React from "react";
-    import Center from "../Common/Center";
-    import Button from '../Common/Button'
-    import Cube from '../../images/Cube.png'
+import Center from "../Common/Center";
+import Button from "../Common/Button";
 
-    const EmptyBox = () => {
-    return (
-        <Center>
-        <>
-        <div className="vertical-content">
-        <img src={Cube} alt='empty cube'/>
-        <p>Start by adding media assets<br/> Click <b>Upload assets</b> to upload your data</p>
-        <Button label="upload assets"/>
-        </div>
-        </>
-        </Center>
-    );
-    };
+const EmptyBox = () => {
+  return (
+    <Center>
+      <div>
+        <img src="images/cube.png" alt="empty cube" />
+        <p>
+          Start by adding media assets
+          <br /> Click <b>Upload assets</b> to upload your data
+        </p>
+        <Button label="Upload assets" />
+      </div>
+    </Center>
+  );
+};
 
-    export default EmptyBox;
+export default EmptyBox;
