--- src/components/Home/Home.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import img from '../../images/commingsoon.png'
-
-function Home() {
-    return (
-        <div className="w-full h-full bg-[#F2EBE1]">
-            <h1>comming soon...</h1>
-        </div>
-    );
-}
-
-export default Home;
