+++ src/components/Navigation/index.tsx
@@ -2,25 +2,17 @@ import { useSignOut } from "react-firebase-hooks/auth";
 import { auth } from "../../helpers/firebase_utils";
 import Button from "../Common/Button";
 import Center from "../Common/Center";
-import s from "./Dashboard.module.css";
+import s from "./Navigation.module.css";
 
-const Dashboard = () => {
+const Navigation = () => {
   const [signOut, loading, error] = useSignOut(auth);
 
   if (error) return <p>Error: {error.message}</p>;
-  if (loading) return <p>Loading from Dashboard...</p>;
+  if (loading) return <p>Loading from Navigation...</p>;
   return (
     <div>
       <Center className="h-28">
-        <Button
-          label="Logout"
-          onClick={async () => {
-            const success = await signOut();
-            if (success) {
-              alert("You are sign out");
-            }
-          }}
-        />
+        <Button label="Create New Page" onClick={() => {}} />
       </Center>
       <div className={s.container}>
         <h1 className={s.title}>PAGES</h1>
@@ -31,8 +23,16 @@ const Dashboard = () => {
           <li>Product Image Gallery</li>
         </ul>
       </div>
+      <Center className="h-28">
+        <Button
+          label="Logout"
+          onClick={async () => {
+            const success = await signOut();
+          }}
+        />
+      </Center>
     </div>
   );
 };
 
-export default Dashboard;
+export default Navigation;
