--- src/components/SocialLogin/Sociallogin.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-function Sociallogin() {
-    return (
-        <div>
-            <ul>
-                <li>Google</li>
-                <li>Facebook</li>
-            </ul>
-        </div>
-    );
-}
-
-export default Sociallogin;
